+++ src/app/ConfirmPage.tsx
@@ -1,4 +1,3 @@
-import BigNumber from "bignumber.js";
 import * as React from "react";
 import classNames from "clsx";
 import { ThanosDAppMetadata } from "@thanos-wallet/dapp/dist/types";
@@ -11,11 +10,6 @@ import {
   ThanosDAppPayload,
   XTZ_ASSET,
   ThanosAccount,
-  ThanosAsset,
-  useSwapData,
-  Swap,
-  useTransfersData,
-  Transfer,
 } from "lib/thanos/front";
 import { useRetryableSWR } from "lib/swr";
 import useSafeState from "lib/ui/useSafeState";
@@ -25,10 +19,8 @@ import Unlock from "app/pages/Unlock";
 import ContentContainer from "app/layouts/ContentContainer";
 import AccountBanner from "app/templates/AccountBanner";
 import NetworkBanner from "app/templates/NetworkBanner";
-import OperationsBanner from "app/templates/OperationsBanner";
 import Balance from "app/templates/Balance";
 import CustomSelect, { OptionRenderProps } from "app/templates/CustomSelect";
-import FormField from "app/atoms/FormField";
 import Logo from "app/atoms/Logo";
 import Identicon from "app/atoms/Identicon";
 import Name from "app/atoms/Name";
@@ -39,16 +31,11 @@ import FormSubmitButton from "app/atoms/FormSubmitButton";
 import FormSecondaryButton from "app/atoms/FormSecondaryButton";
 import ConfirmLedgerOverlay from "app/atoms/ConfirmLedgerOverlay";
 import HashShortView from "app/atoms/HashShortView";
-import { ReactComponent as ArrowRightIcon } from "app/icons/arrow-right.svg";
 import { ReactComponent as ComponentIcon } from "app/icons/component.svg";
 import { ReactComponent as OkIcon } from "app/icons/ok.svg";
 import { ReactComponent as LayersIcon } from "app/icons/layers.svg";
-import { ReactComponent as EyeIcon } from "app/icons/eye.svg";
-import { ReactComponent as CodeAltIcon } from "app/icons/code-alt.svg";
-import { ReactComponent as SwapIcon } from "app/icons/swap.svg";
-import { ReactComponent as SendIcon } from "app/icons/send.svg";
 import DAppLogo from "app/templates/DAppLogo";
-import { getAssetIconUrl } from "app/defaults";
+import OperationView from "app/ConfirmPage/OperationView";
 
 const ConfirmPage: React.FC = () => {
   const { ready } = useThanosClient();
@@ -489,397 +476,6 @@ const AccountOptionContentHOC = (networkRpc: string) => {
     </>
   ));
 };
-
-type OperationViewProps = {
-  payload: ThanosDAppPayload;
-};
-
-const OperationView: React.FC<OperationViewProps> = (props) => {
-  const { payload } = props;
-  const contentToParse = React.useMemo(() => {
-    switch (payload.type) {
-      case "confirm_operations":
-        return payload.opParams;
-      case "sign":
-        return payload.preview;
-      default:
-        return [];
-    }
-  }, [payload]);
-  const swapData = useSwapData(contentToParse);
-  const transfersData = useTransfersData(contentToParse);
-  console.log(swapData, transfersData);
-
-  const signPayloadFormats = React.useMemo(() => {
-    const previewFormat = {
-      key: "preview",
-      name: t("preview"),
-      Icon: EyeIcon,
-    };
-    const prettyViewFormats = [
-      swapData
-        ? {
-            key: "swap",
-            name: t("swapNoun"),
-            Icon: SwapIcon,
-          }
-        : undefined,
-      transfersData.length > 0 && !swapData
-        ? {
-            key: "transfers",
-            name: t("transfers"),
-            Icon: SendIcon,
-          }
-        : undefined,
-    ].filter((item): item is ViewsSwitcherItemProps => !!item);
-
-    if (payload.type === "confirm_operations") {
-      return [previewFormat, ...prettyViewFormats];
-    }
-
-    if (payload.type === "connect") {
-      return [];
-    }
-
-    return [
-      previewFormat,
-      {
-        key: "raw",
-        name: t("raw"),
-        Icon: CodeAltIcon,
-      },
-      ...prettyViewFormats,
-    ];
-  }, [payload.type, swapData, transfersData]);
-
-  const [spFormat, setSpFormat] = React.useState(signPayloadFormats[0]);
-
-  if (payload.type === "sign" && payload.preview) {
-    return (
-      <div className="flex flex-col w-full">
-        <h2
-          className={classNames("mb-4", "leading-tight", "flex items-center")}
-        >
-          <T id="payloadToSign">
-            {(message) => (
-              <span
-                className={classNames(
-                  "mr-2",
-                  "text-base font-semibold text-gray-700"
-                )}
-              >
-                {message}
-              </span>
-            )}
-          </T>
-
-          <div className="flex-1" />
-
-          <ViewsSwitcher
-            activeItem={spFormat}
-            items={signPayloadFormats}
-            onChange={setSpFormat}
-          />
-        </h2>
-
-        <OperationsBanner
-          opParams={payload.preview}
-          label={null}
-          className={classNames(spFormat.key !== "preview" && "hidden")}
-        />
-
-        <RawPayloadView
-          payload={payload.payload}
-          className={classNames(spFormat.key !== "raw" && "hidden")}
-        />
-
-        <div className={classNames(spFormat.key !== "swap" && "hidden")}>
-          <SwapView swap={swapData} />
-        </div>
-
-        <div className={classNames(spFormat.key !== "transfers" && "hidden")}>
-          <TransfersView transfers={transfersData} />
-        </div>
-      </div>
-    );
-  }
-
-  if (payload.type === "sign") {
-    return (
-      <RawPayloadView
-        label={t("payloadToSign")}
-        payload={payload.payload}
-        className="mb-2"
-      />
-    );
-  }
-
-  if (payload.type === "confirm_operations") {
-    return (
-      <div className="flex flex-col w-full">
-        <h2
-          className={classNames("mb-4", "leading-tight", "flex items-center")}
-        >
-          <span
-            className={classNames(
-              "mr-2",
-              "text-base font-semibold text-gray-700"
-            )}
-          >
-            <T id="operations" />
-          </span>
-
-          <div className="flex-1" />
-
-          {signPayloadFormats.length > 1 && (
-            <ViewsSwitcher
-              activeItem={spFormat}
-              items={signPayloadFormats}
-              onChange={setSpFormat}
-            />
-          )}
-        </h2>
-
-        <OperationsBanner
-          opParams={payload.opParams}
-          className={classNames(spFormat.key !== "preview" && "hidden")}
-          label={null}
-        />
-
-        <div className={classNames(spFormat.key !== "swap" && "hidden")}>
-          <SwapView swap={swapData} />
-        </div>
-
-        <div className={classNames(spFormat.key !== "transfers" && "hidden")}>
-          <TransfersView transfers={transfersData} />
-        </div>
-      </div>
-    );
-  }
-
-  return null;
-};
-
-type SwapViewProps = {
-  swap?: Swap;
-};
-
-const SwapView: React.FC<SwapViewProps> = (props) => {
-  const { swap } = props;
-
-  if (!swap) {
-    return null;
-  }
-  return (
-    <>
-      <p className="mb-2 text-base font-semibold text-gray-700 text-center">
-        <T id="swapNoun" />
-      </p>
-      <div className="flex flex-no-wrap">
-        <SwapAssetView
-          amount={
-            swap.type === "xtzToToken" ? swap.xtzAmount : swap.tokenInAmount
-          }
-          asset={swap.type === "xtzToToken" ? XTZ_ASSET : swap.tokenIn}
-        />
-        <div className="flex-none flex items-center mx-2">
-          <ArrowRightIcon className="w-6 h-auto stroke-current" title="to" />
-        </div>
-        <SwapAssetView
-          amount={
-            swap.type === "tokenToXtz" ? swap.xtzAmount : swap.tokenOutAmount
-          }
-          asset={swap.type === "tokenToXtz" ? XTZ_ASSET : swap.tokenOut}
-        />
-      </div>
-    </>
-  );
-};
-
-type SwapAssetViewProps = {
-  asset: ThanosAsset | string;
-  amount: BigNumber;
-};
-
-const SwapAssetView: React.FC<SwapAssetViewProps> = (props) => {
-  const { asset, amount } = props;
-
-  return (
-    <div className="flex-1 flex flex-col items-center justify-center">
-      {typeof asset === "string" ? (
-        <span className="text-base text-gray-700">
-          <Money>{amount}</Money> (
-          <T
-            id="someUnknownToken"
-            substitutions={<HashShortView hash={asset} />}
-          />
-          )
-        </span>
-      ) : (
-        <>
-          <img
-            className="h-8 w-auto mb-2"
-            alt={asset.symbol}
-            src={getAssetIconUrl(asset)}
-          />
-          <span className="text-xl text-gray-700">
-            <Money>{amount}</Money>{" "}
-            <span className="text-base">{asset.symbol}</span>
-          </span>
-        </>
-      )}
-    </div>
-  );
-};
-
-type TransfersViewProps = {
-  transfers?: Transfer[];
-};
-
-const TransfersView = React.memo<TransfersViewProps>((props) => {
-  const { transfers } = props;
-
-  if (!transfers) {
-    return null;
-  }
-
-  return (
-    <>
-      {transfers.map((transfer, index) => (
-        <React.Fragment key={index}>
-          <div className="mb-2 flex flex-no-wrap">
-            <TransferSideView accountPkh={transfer.from} />
-            <div className="flex-none flex items-center mx-2">
-              <ArrowRightIcon
-                className="w-6 h-auto stroke-current"
-                title="to"
-              />
-            </div>
-            <TransferSideView accountPkh={transfer.to} />
-          </div>
-          <p className="text-xl mb-2 text-center text-gray-700">
-            <Money>{transfer.amount}</Money>{" "}
-            <span className="text-base">
-              {transfer.type === "transferToken" &&
-                (typeof transfer.token === "string" ? (
-                  <>
-                    (
-                    <T
-                      id="someUnknownToken"
-                      substitutions={<HashShortView hash={transfer.token} />}
-                    />
-                    )
-                  </>
-                ) : (
-                  transfer.token.symbol
-                ))}
-              {transfer.type === "transferXTZ" && "XTZ"}
-            </span>
-          </p>
-        </React.Fragment>
-      ))}
-    </>
-  );
-});
-
-type TransferSideViewProps = {
-  accountPkh: string;
-};
-
-const TransferSideView: React.FC<TransferSideViewProps> = (props) => {
-  const { accountPkh } = props;
-
-  return (
-    <div className="flex-1 flex flex-col items-center justify-center">
-      <Identicon className="mb-2" hash={accountPkh} size={32} />
-      <span>
-        <HashShortView hash={accountPkh} />
-      </span>
-    </div>
-  );
-};
-
-type RawPayloadViewProps = {
-  label?: string;
-  payload: string;
-  className?: string;
-};
-
-const RawPayloadView = React.memo(
-  ({ className, payload, label }: RawPayloadViewProps) => (
-    <FormField
-      textarea
-      rows={6}
-      id="sign-payload"
-      label={label}
-      value={payload}
-      spellCheck={false}
-      readOnly
-      className={className}
-      style={{
-        resize: "none",
-      }}
-    />
-  )
-);
-
-type ViewsSwitcherItemProps = {
-  Icon: React.FunctionComponent<React.SVGProps<SVGSVGElement>>;
-  key: string;
-  name: string;
-};
-
-type ViewsSwitcherProps = {
-  activeItem: ViewsSwitcherItemProps;
-  items: ViewsSwitcherItemProps[];
-  onChange: (item: ViewsSwitcherItemProps) => void;
-};
-
-const ViewsSwitcher = React.memo(
-  ({ activeItem, items, onChange }: ViewsSwitcherProps) => (
-    <div className={classNames("flex items-center")}>
-      {items.map((spf, i, arr) => {
-        const first = i === 0;
-        const last = i === arr.length - 1;
-        const selected = activeItem.key === spf.key;
-        const handleClick = () => onChange(spf);
-
-        return (
-          <button
-            key={spf.key}
-            className={classNames(
-              (() => {
-                switch (true) {
-                  case first:
-                    return classNames("rounded rounded-r-none", "border");
-
-                  case last:
-                    return classNames(
-                      "rounded rounded-l-none",
-                      "border border-l-0"
-                    );
-
-                  default:
-                    return "border border-l-0";
-                }
-              })(),
-              selected && "bg-gray-100",
-              "px-2 py-1",
-              "text-xs text-gray-600",
-              "flex items-center"
-            )}
-            onClick={handleClick}
-          >
-            <spf.Icon
-              className={classNames("h-4 w-auto mr-1", "stroke-current")}
-            />
-            {spf.name}
-          </button>
-        );
-      })}
-    </div>
-  )
-);
 
 type ConnectBannerProps = {
   type: "connect" | "confirm_operations";
