+++ src/lib/thanos/front/index.ts
@@ -13,5 +13,4 @@ export * from "lib/thanos/front/balance";
 export * from "lib/thanos/front/baking";
 export * from "lib/thanos/front/tokens";
 export * from "lib/thanos/front/assets";
-export * from "lib/thanos/front/swap";
-export * from "lib/thanos/front/transfer";
+export * from "lib/thanos/front/expenses";
